Address Review Feedback: Remove zip step from bin extract and output raw binaries in llb.State Move bin extract tests under testLinuxDistro Add more descriptive test error messages
